fix 102 fib

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,22 +1,20 @@
 #include <stdio.h>
 /**
-* fib - function to calculate fibonacci
-* @n: the number
-* Return: fib of n
+* fib50 - function to calculate fibonacci
 */
 
 void fib50(void)
 {
-	long long n1 = 1;
-	long long n2 = 2;
-	long long sum = n1 + n2;
+	long n1 = 1;
+	long n2 = 2;
+	long sum = n1 + n2;
 	int i = 0;
 
-	printf("%lld, ", n1);
-	printf("%lld, ", n2);
+	printf("%ld, ", n1);
+	printf("%ld, ", n2);
 	while (i < 50 - 2)
 	{
-		printf("%lld, ", sum);
+		printf("%ld, ", sum);
 		n1 = n2;
 		n2 = sum;
 		sum = n2 + n1;
